requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
pyotp
pydantic
requests
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
email-validator
aiosmtplib 